refactor: move scrolling related code into separate context
