+++ ui/src/notebooks/components/Notebook.tsx
@@ -2,6 +2,7 @@ import React, {FC} from 'react'
 
 import {Page} from '@influxdata/clockface'
 import {NotebookProvider} from 'src/notebooks/context/notebook'
+import {ScrollProvider} from 'src/notebooks/context/scroll'
 import Header from 'src/notebooks/components/header'
 import PipeList from 'src/notebooks/components/PipeList'
 import MiniMap from 'src/notebooks/components/minimap/MiniMap'
@@ -14,19 +15,21 @@ import 'src/notebooks/style.scss'
 const NotebookPage: FC = () => {
   return (
     <NotebookProvider>
-      <Page titleTag="Notebook">
-        <Header />
-        <Page.Contents
-          fullWidth={true}
-          scrollable={false}
-          className="notebook-page"
-        >
-          <div className="notebook">
-            <MiniMap />
-            <PipeList />
-          </div>
-        </Page.Contents>
-      </Page>
+      <ScrollProvider>
+        <Page titleTag="Notebook">
+          <Header />
+          <Page.Contents
+            fullWidth={true}
+            scrollable={false}
+            className="notebook-page"
+          >
+            <div className="notebook">
+              <MiniMap />
+              <PipeList />
+            </div>
+          </Page.Contents>
+        </Page>
+      </ScrollProvider>
     </NotebookProvider>
   )
 }
