+++ ui/src/notebooks/components/PipeList.tsx
@@ -3,6 +3,7 @@ import React, {FC, useContext, useCallback} from 'react'
 
 // Contexts
 import {NotebookContext} from 'src/notebooks/context/notebook'
+import {ScrollContext} from 'src/notebooks/context/scroll'
 
 // Components
 import NotebookPipe from 'src/notebooks/components/NotebookPipe'
@@ -10,9 +11,8 @@ import EmptyPipeList from 'src/notebooks/components/EmptyPipeList'
 import {DapperScrollbars} from '@influxdata/clockface'
 
 const PipeList: FC = () => {
-  const {id, pipes, updatePipe, listScrollPosition} = useContext(
-    NotebookContext
-  )
+  const {id, pipes, updatePipe} = useContext(NotebookContext)
+  const {scrollPosition} = useContext(ScrollContext)
   const update = useCallback(updatePipe, [id])
 
   if (!pipes.length) {
@@ -35,7 +35,7 @@ const PipeList: FC = () => {
       className="notebook-main"
       autoHide={true}
       noScrollX={true}
-      scrollTop={listScrollPosition}
+      scrollTop={scrollPosition}
     >
       {_pipes}
     </DapperScrollbars>
